+++ example/android/app/src/main/kotlin/com/example/button_loading_fx_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.button_loading_fx_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()